Set up URP 2d renderer and pixel perfect camera, and add idle animation
This creates a new URP 2d renderer in the project directory and sets it as the default graphics renderer in the project settings as crisp and pixel perfect rendering is essential in this project. It also adds pixel perfect component to the camera as well as adding the new idle animation to the character alongside a default sprite instead of a square.